refactor(api): split group.service.ts into services/group/ (api-jhyt)
