+++ apps/api/src/routes/groups/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { archiveGroup } from "../../services/group.service.js";
+import { archiveGroup } from "../../services/group/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
