+++ apps/api/src/routes/groups/reorder.ts
@@ -7,7 +7,7 @@ import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { reorderGroups } from "../../services/group.service.js";
+import { reorderGroups } from "../../services/group/structure.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
