+++ apps/api/src/services/group/update.ts
@@ -1,0 +1,16 @@
+import { groupHierarchy } from "./internal.js";
+
+import type { GroupResult } from "./internal.js";
+import type { AuditWriter } from "../../lib/audit-writer.js";
+import type { AuthContext } from "../../lib/auth-context.js";
+import type { GroupId, SystemId } from "@pluralscape/types";
+import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
+
+export const updateGroup: (
+  db: PostgresJsDatabase,
+  systemId: SystemId,
+  entityId: GroupId,
+  params: unknown,
+  auth: AuthContext,
+  audit: AuditWriter,
+) => Promise<GroupResult> = groupHierarchy.update;
